+++ decitala/fragment.py
@@ -15,7 +15,7 @@ import numpy as np
 import os
 
 from collections import Counter
-from functools import cache
+from functools import lru_cache
 
 from music21 import converter
 from music21 import note
@@ -258,7 +258,7 @@ class GeneralFragment:
 		return utils.ql_array_to_greek_diacritics(self.ql_array())
 
 	@property
-	@cache  # Caching *extremely* useful for cost function in path-finding.
+	@lru_cache(maxsize=None)  # Caching *extremely* useful for cost function in path-finding.
 	def num_onsets(self):
 		"""
 		:return: The number of onsets in the fragment.