+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All important changes to the decitala package will be documented here.
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and the project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [v0.10.1] ???
+### Fixed
+- The `utils.find_possible_superdivisions` function was including itself as a possible superdivision. This was a bit confusing, so I added an `include_self` parameter. 
 
 ## [v0.10.0] April 29, 2021
 ### Added